squads/piranha-dev/projects/piranha-leads/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
lxml==5.2.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.3
